add all services button
