+++ messages/en.json
@@ -281,7 +281,8 @@
         "titleSEO": "Website development services for business - modern web design and development",
         "descriptionSEO": "Turnkey web development of websites: convenient structure, modern design and integration of functions. We work with individual requirements of each client.",
         "ctaText": "Get a comprehensive consultation with our manager if you have any questions or doubts",
-        "details": "More"
+        "details": "More",
+        "allService": "All services"
     },
     "Christmas": {
         "titleSEO": "Merry Christmas!",
