+++ messages/pl.json
@@ -282,7 +282,8 @@
         "titleSEO": "Usługi tworzenia stron internetowych dla biznesu - nowoczesne projektowanie i tworzenie stron internetowych",
         "descriptionSEO": "Tworzenie stron internetowych pod klucz: przyjazna dla użytkownika struktura, nowoczesny design i integracja funkcji. Pracujemy z indywidualnymi wymaganiami każdego klienta.",
         "ctaText": "Uzyskaj kompleksową konsultację z naszym menedżerem, jeśli masz jakiekolwiek pytania lub wątpliwości",
-        "details": "Więcej"
+        "details": "Więcej",
+        "allService": "Wszystkie usługi"
     },
     "Christmas": {
         "titleSEO": "Wesołych Świąt i Szczęśliwego Nowego Roku!",
