+++ messages/ua.json
@@ -278,7 +278,8 @@
         "titleSEO": "Послуги створення сайтів для бізнесу — сучасний вебдизайн і розробка",
         "descriptionSEO": "Веброзробка сайтів під ключ: зручна структура, сучасний дизайн і інтеграція функцій. Працюємо з індивідуальними вимогами кожного клієнта.",
         "ctaText": "Отримайте вичерпну консультацію з нашим менеджером, якщо маєте якісь питання або сумніви",
-        "details": "Детальніше"
+        "details": "Детальніше",
+        "allService": "Всі послуги"
     },
     "Christmas": {
         "titleSEO": "З Різдвом та Новим Роком!",
